verificação se tem pets no tutor para poder apagar

+++ src/service/tutor.service.ts
@@ -1,6 +1,9 @@
 import { Request, Response } from "express";
 import Tutor from "../model/tutor.model";
 
+import { _Pet } from "../model/pet.model";
+import { _Tutor } from "../model/tutor.model";
+
 export async function createTutor(req: Request, res: Response) {
   const { name, phone, email, date_of_birth, zip_code, pets } = req.body;
 
@@ -41,17 +44,29 @@ export async function getById(req: Request, res: Response) {
   }
 }
 
-export async function deleteById(req: Request, res: Response) {
+export async function deleteTutor(req: Request, res: Response) {
   const id = req.params.id;
+
   try {
-    const tutor = await Tutor.deleteOne({ _id: id });
-    res.status(204).json({ tutor, message: "Tutor apagado" });
+    const tutor = await Tutor.findById(id);
+    if (!tutor) {
+      return res.status(404).json({ error: 'Tutor não encontrado' });
+    }
+
+    if (tutor.pets.length > 0) {
+      return res.status(400).json({ error: 'Não é possível excluir o tutor porque ele possui pets associados.' });
+    }
+
+    await Tutor.deleteOne({ _id: id });
+
+    res.status(204).json({ message: 'Tutor apagado' });
   } catch (error) {
     res.status(500).json({ error: error });
   }
 }
 
-export async function updateById(req: Request, res: Response) {
+
+export async function updateTutor(req: Request, res: Response) {
   const id = req.params.id;
 
   const { name, phone, email, date_of_birth, zip_code, pets } = req.body;
